+++ CHANGELOG.md
@@ -15,10 +15,15 @@ once the API is stable.
 
   A function to determine of a string is a valid channel name.
 
+- ADDED: `parsers.kwargs_to_kwargs`
+
+  A handler decorator that takes a parser, and passes the resulting dictionary
+  to the handler as kwargs. The parser should accept all kwargs.
+
 - ADDED: `parsers.message_to_kwargs`
 
   A handler decorator that takes a parser, and passes the resulting dictionary
-  to the handler as kwargs.
+  to the handler as kwargs. The parser should accept only a `message` kwarg.
 
 - ADDED: `parsers.nick`.